% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_identical.R
\name{all_identical}
\alias{all_identical}
\title{Test if all elements of a vector are identical}
\usage{
all_identical(x, empty_value = FALSE)
}
\arguments{
\item{x}{vector to be tested}

\item{empty_value}{Value to return if all_identical is called on a vector of length 0}
}
\value{
TRUE/FALSE
}
\description{
Test if all elements of a vector are identical
}
\examples{

all_identical(c(1,2,3))
all_identical(c(1,1,1))

}
