% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_chop.R
\name{chop}
\alias{chop}
\title{Split up a vector into equal sized chunks}
\usage{
chop(x, interval)
}
\arguments{
\item{input}{An input vector.}

\item{by}{The length of the resulting vectors.}
}
\description{
https://gist.github.com/sckott/4632735
}
\examples{
vec <- c("a","b","d","e","f","g","h")
chop(vec, 3)
}

