% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{condition}
\alias{condition}
\title{Condition constructor}
\usage{
condition(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{Subclass to assign to the condition}

\item{message}{message to be passed to the condition}

\item{call}{call passed on to the conditon}

\item{...}{further list elements to be passed on to the resulting object}
}
\value{
a condition object
}
\description{
A constructur function for conditions, taken rom
http://adv-r.had.co.nz/beyond-exception-handling.html
}

