% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_db2_utils.R
\name{db2ExistsTable}
\alias{db2ExistsTable}
\title{Check if db2 table exists}
\usage{
db2ExistsTable(con, tname)
}
\arguments{
\item{con}{an \code{\link{RODBC}} connection object}

\item{tname}{name of the table}
}
\value{
TRUE/FALSE if table exists
}
\description{
Check if db2 table exists
}

