% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_affix.R
\name{df_affix}
\alias{df_affix}
\alias{df_braces}
\alias{df_brackets}
\alias{df_parenthesis}
\title{Paste a suffix and/or prefix to each column of a data.frame}
\usage{
df_affix(dat, prefix, suffix)

df_parenthesis(dat)

df_brackets(dat)

df_braces(dat)
}
\arguments{
\item{dat}{a data.frame}

\item{prefix}{to be pasted before each elemtn}

\item{suffix}{}
}
\value{
a \code{data.frame} (with only character columns)
}
\description{
Paste a suffix and/or prefix to each column of a data.frame
}
\examples{
dat <- data.frame(
  a = c(1, 2),
  b = c('a', 'b'),
  stringsAsFactors = FALSE
)

df_affix(dat, "prefix-", "-suffix" )
df_parenthesis(dat)
df_brackets(dat)

}

