% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_drop_cols.R
\name{df_drop_cols}
\alias{df_drop_cols}
\title{Drop columns of a data.frame}
\usage{
df_drop_cols(dat, drop, allow_partial = TRUE)
}
\arguments{
\item{dat}{a data.frame}

\item{drop}{(character vector) names of the columns to be droped}

\item{allow_partial}{(logical vector) If set to FALSE an error will be thrown
if not all \code{drop} columns are present in \code{dat}}
}
\description{
Drop columns of a data.frame
}

