% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_format.R
\name{df_format_num}
\alias{df_format_num}
\title{Format numeric columns of a data.frame}
\usage{
df_format_num(dat, integer_as_numeric = TRUE, ...)
}
\arguments{
\item{dat}{a \code{data.frame} (or something that can be coearced with \code{as.data.frame})}

\item{integer_as_numeric}{Logical. Should integer columns be formatted the same way as numeric columns?}

\item{...}{Params passed on to \code{format}}
}
\description{
This function applies formatting to each numeric column of a data.frame
For details refer to the documentation of \code{\link{format}}
}
\seealso{
df_format
}
