% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_round.R
\name{df_signif}
\alias{df_signif}
\title{Round columns of a data frame to a specified number of significant digits.}
\usage{
df_signif(dat, digits = 0)
}
\arguments{
\item{digits}{}
}
\description{
Rounds the values in the numeric columns of a data.frame to the specified number of significant digits.
}
\seealso{
signif
}

