% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_trimws.R
\name{df_trimws}
\alias{df_trimws}
\title{Remove whitespace from all character (and factor) columns of a data.frame}
\usage{
df_trimws(dat, process_factors = FALSE)
}
\arguments{
\item{dat}{a data.frame}

\item{process_factors}{wheter or not factor labels should also be processed}
}
\value{
a data.frame
}
\description{
Remove whitespace from all character (and factor) columns of a data.frame
}

