% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_typecast_cols.R
\name{df_typecast_cols}
\alias{df_typecast_cols}
\title{Typecast columns of a data.frame by name}
\usage{
df_typecast_cols(dat, conv = list())
}
\arguments{
\item{dat}{a data.frame}

\item{conv}{a list of the form list(COLNAME = 'coltype')}
}
\value{
a data.frame with typecasted columns
}
\description{
Use with care, will introduce NAs for some conversion attempts
}
\examples{
dat <- data.frame(foo = c('5', '6', '5'),
                  bar = factor(c('a', 'b', 'c')),
                  stringsAsFactors = FALSE)
str(dat)

res <- typecast_cols(dat, list(foo = 'numeric',
                              bar = 'character'))

str(dat)
}

