% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ftp.R
\name{download_ftp}
\alias{download_ftp}
\title{Download files via FTP}
\usage{
download_ftp(.file, .outdir = ".", .creds = NULL, .server,
  .overwrite = FALSE, .mode = "ascii", ...)
}
\arguments{
\item{.file}{name of the file (or files) to download}

\item{.outdir}{local dir to download file to}

\item{.creds}{login credentials (see \code{\link{ui_credentials}})}

\item{.server}{download server adress}

\item{.overwrite}{should local files be overwritten?}

\item{.mode}{"binary" or "ascii"}

\item{...}{}
}
\description{
Calls command line FTP client, only tested on windows, you should probably
rather use the RCurl package for this task
}
