% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fetch_host_file.R
\name{download_host_file}
\alias{download_host_file}
\title{Fetch files from host via FTP}
\usage{
download_host_file(file, outdir = ".", creds = ftp_creds,
  server = "mfstat01", overwrite = FALSE, ...)
}
\arguments{
\item{file}{filename on host}

\item{outdir}{local destination directory}

\item{creds}{login credentials for host}

\item{overwrite}{should existing files be overwritten?}

\item{...}{params passed to download_ftp}
}
\description{
Fetch files from host via FTP
}

