% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_drop_cols.R
\name{drop_if_exists}
\alias{drop_if_exists}
\title{Drop a column from a data.frame if it exists}
\usage{
drop_if_exists(dat, drop)
}
\arguments{
\item{dat}{data.frame}
}
\value{
a data.frame without the columns specified in drop
}
\description{
Drop a column from a data.frame if it exists
}

