% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_filename_from_path.R
\name{extract_filename_from_path}
\alias{extract_filename_from_path}
\title{Extracts the filename from a path}
\usage{
extract_filename_from_path(x, fsep = NULL, ext = TRUE)
}
\arguments{
\item{x}{a character vector containing a path to a file}

\item{ext}{Wheter or not the result should include the file extension or not}
}
\value{
The filename
}
\description{
Extracts the filename from a path
}
\examples{

 x <- c("C:/path/to/test.file.ext")

 extract_filename_from_path(x)
 extract_filename_from_path(x, ext = FALSE)
}

