% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{fct_recode2}
\alias{fct_recode2}
\title{Change factor levels according to named character vector}
\usage{
fct_recode2(x, rec)
}
\arguments{
\item{rec}{}
}
\description{
This is just a (to me) slightly more convenient interface to
forcats::fct_recode.
}
