% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_host_db2.R, R/stat_fetch_host_db2.R
\name{fetch_host_db2}
\alias{fetch_host_db2}
\title{Fetch db2 tables from host}
\usage{
fetch_host_db2(table, con = RODBC::odbcConnect(dsn = gvk_secrets["dsn"], uid =
  gvk_secrets["uid"], pwd = gvk_secrets["pwd"]))

fetch_host_db2(table, con = RODBC::odbcConnect(dsn = gvk_secrets["dsn"], uid =
  gvk_secrets["uid"], pwd = gvk_secrets["pwd"]))
}
\arguments{
\item{table}{name of the table to be retrieved}

\item{con}{an ROBC connection object}

\item{table}{name of the table to be retrieved}

\item{con}{an ROBC connection object}
}
\description{
Fetch db2 tables from host

Fetch db2 tables from host
}
\section{Tips}{


Fetch 'sysibm.systables' to get infos about all tables on db2 server (warning: big query)



Fetch 'sysibm.systables' to get infos about all tables on db2 server (warning: big query)
}

