% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fetch_host_file.R
\name{generate_yq_filename}
\alias{generate_yq_filename}
\title{Title}
\usage{
generate_yq_filename(year, quarter = NULL, mask = "example_\%s-Q\%s",
  reverse = FALSE)
}
\arguments{
\item{reverse}{}
}
\description{
Title
}

