% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Extract legend from a ggplot 2 object}
\usage{
get_legend(dat)
}
\arguments{
\item{dat}{a ggplot2 object}
}
\value{
a legend or an empty grob
}
\description{
source: http://stackoverflow.com/questions/13649473/add-a-common-legend-for-combined-ggplots
}

