% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datetime.R
\name{get_year}
\alias{get_year}
\alias{get_quarter}
\alias{get_month}
\alias{get_month_first}
\title{Get Year, Quarter or Month of a POSIXt object}
\usage{
get_year(x)

get_quarter(x)

get_month(x)

get_month_first(x)
}
\arguments{
\item{x}{A POSIXt object}
}
\value{
A character repressentation of the year of x
}
\description{
Get Year, Quarter or Month of a POSIXt object
}
\examples{

get_year(Sys.time())
}
