% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ggplotify_spdf.R
\name{ggplotify_spdf}
\alias{ggplotify_spdf}
\title{Prepare SPDF for ggplot2}
\usage{
ggplotify_spdf(sp, region)
}
\arguments{
\item{sp}{a sp object}

\item{region}{a region identifier}
}
\value{
a data.frame
}
\description{
Prepare SPDF for ggplot2
}

