% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_shortcuts.R
\name{\%==\%}
\alias{\%==\%}
\title{Equal and not not NA}
\usage{
a \%==\% b
}
\description{
Works like ==, just that it treats NAs like normal values:
}
\details{
NA %==% NA returns TRUE instead of NA
  1  %==% NA returns FALSE instead of NA
}

