% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_numbers.R
\name{human_numbers}
\alias{human_numbers}
\alias{human_gbp,}
\alias{human_usd,}
\alias{human_euro,}
\alias{human_num,}
\alias{human_num_de,}
\alias{human_si}
\alias{human_gbp}
\alias{human_usd}
\alias{human_euro}
\alias{human_num}
\alias{human_num_de}
\alias{human_si}
\title{Human Numbers}
\usage{
human_numbers(x, symbol = "", pots = c(k = 1000, m = 1e+06, b = 1e+09),
  big_mark = ",")

human_gbp(x)

human_usd(x)

human_euro(x)

human_num(x)

human_num_de(x)

human_si(x)
}
\arguments{
\item{x}{a numeric vector}

\item{symbol}{a prefix symbol, useful for currencies}

\item{pots}{A named numeric vector of rounding levels and associated shorthand}
}
\value{
a character vector the same length as the input vector
}
\description{
Format numbers so they're legible for humans. Use this in ggplot for labels
where you might use the comma or percent functions from the Scales package.
}
\details{
Comes with convenience functions for currencies, SI symbols and and german
axis labels.

Inspired by: https://github.com/1R151-1/R/blob/master/ggplot2_formatter.r
}
\examples{
human_numbers(c(1000000 , 1500000, 10000000000))
human_numbers(c(1.200000e+05, -2.154660e+05, 2.387790e+05, 4.343500e+04 ,5.648675e+12), "$")


x <- data.frame(
  x = letters[sample(1:5, 1000, replace = TRUE)],
  y = c(runif(1000, min = -1e12, max = 1e12))
)

p <- ggplot(x, aes(x = x, y = y)) + geom_point()
p + scale_y_continuous(labels = human_num)
p + scale_y_continuous(labels = human_euro)
}
