% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment.R
\name{increment}
\alias{increment}
\alias{increment.Quarter}
\title{Increment functions for various self-defined objects}
\usage{
increment(x, inc = 1)

\method{increment}{Quarter}(x, inc)
}
\arguments{
\item{x}{object to increment}

\item{inc}{Value by which to increment (usually integer)}
}
\value{
An object of the same type increment by inc
}
\description{
Function to increment objects
}

