% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_class}
\alias{\%assert_class\%}
\alias{\%is_class\%}
\alias{assert_class}
\alias{is_class}
\title{Check if object is of a certain class}
\usage{
is_class(dat, class)

assert_class(dat, class)

dat \%assert_class\% class

dat \%is_class\% class
}
\arguments{
\item{dat}{any R object}

\item{class}{the class to be checked for}
}
\value{
is_class returns TRUE/FALSE, assert_class returns TRUE or fails with
        an error message.
}
\description{
These functions are desinged to be used in combination with the assertthat
pacakge
}
\details{
is_class returns TRUE/FALSE. It comes with a on_failure function and
is designed to be used in conjunction with the assertthat package.
assert_class() and its infix version %assert_class% fail with an error message
}
\examples{

x = data.frame()
is.class(x, 'data.frame')
}

