% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_col_classes}
\alias{is_col_classes}
\title{Check for column classes}
\usage{
is_col_classes(dat, classes, method = "identical")
}
\arguments{
\item{dat}{a data.frame or list}

\item{classes}{a list of column classes. Its names must match
the names of dat exactly (see example)}

\item{method}{if \code{all}, ensure that all columns named in \code{classes} are present in \code{dat},
if \code{any}, ensure that any of the  columns named in \code{classes} are present in \code{dat},
if \code{identical}, ensure that the names of dat and classes are identical}
}
\description{
Compares the column classes of a data.frame with
}
