% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_edge_pairs.R
\name{is_potentially_sortable}
\alias{is_potentially_sortable}
\title{Is potentially sortable}
\usage{
is_potentially_sortable(p, c, n)
}
\value{
TRUE if there is a chance that the edge pairs could be sorted
}
\description{
FALSE gurantees that not all edge pairs are sortable (but TRUE does
not gurantee that they are)
}

