% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_assertions.R
\name{is_urs}
\alias{is_urs}
\title{Check if all elements of a vector are valid URS}
\usage{
is_urs(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
True if object is of the desired class
}
\description{
Check if all elements of a vector are valid URS
}

