% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kv_range.R
\name{kv_range}
\alias{kv_range}
\title{Key_value range}
\usage{
kv_range(dat, fromcol = "from", tocol = "to")
}
\arguments{
\item{dat}{A data frame. Should contain at leas 4 columns: a key column,
a value column, a from and a to column}

\item{from}{Column of the above data.frame that contains the from values}

\item{to}{Column of the above data.frame that contains the to values}
}
\description{
Key_value range
}

