% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_by_class.R
\name{lapply_by_class}
\alias{lapply_by_class}
\title{Lapply by class}
\usage{
lapply_by_class(x, fun, classes, ...)
}
\arguments{
\item{...}{}
}
\description{
Same as \code{\link{lapply}}, just that it only applies fun to elements of a specified
class.
}
