% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{explorer}
\alias{explorer}
\alias{excel}
\title{Launch explorer}
\usage{
explorer(x)

excel(x,
  excel_path = "\\"C:/Program Files (x86)/Microsoft Office/Office14/EXCEL.EXE\\"")
}
\arguments{
\item{x}{}

\item{x}{}
}
\description{
Launch explorer

Launch excel
}
