% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cache.R
\name{load_cache}
\alias{load_cache}
\title{Load object from hard disk chache (of a package)}
\usage{
load_cache(..., pkg = ".", subdir, envir = globalenv())
}
\arguments{
\item{...}{R objects to load from cache dir, usually /inst/chache}

\item{envir}{target environment}

\item{package}{the current package}
}
\description{
Load object from hard disk chache (of a package)
}
\section{Side effects}{

Loads an R object from a chache dir in the current package
}

\seealso{
\code{\link{save_cache}}.
}
