% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{load_rda}
\alias{load_rda}
\title{Title}
\usage{
load_rda(infile)
}
\arguments{
\item{infile}{}
}
\description{
http://stackoverflow.com/questions/5577221/how-can-i-load-an-object-into-a-variable-name-that-i-specify-from-an-r-data-file
}

