% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_test.R
\name{load_test}
\alias{load_test}
\title{Load object from hard disk chache}
\usage{
load_test(..., pkg = ".", subdir, envir = globalenv())
}
\arguments{
\item{...}{A single r object to load from /tests/testthat/test_data/(subdir)}

\item{package}{the current package}
}
\description{
Load object from hard disk chache
}
\section{Side effects}{

Loads an R object from a test_data dir in the current package
}
\seealso{
\code{\link{save_cache}}.
}

