% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_frequent.R
\name{most_frequent}
\alias{most_frequent}
\title{Get the n most frequent values of a vector}
\usage{
most_frequent(x, n = 1)
}
\arguments{
\item{x}{A vector}

\item{x}{}
}
\value{
a character vector of the n most frequent elements
}
\description{
Get the n most frequent values of a vector
}
