% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_shortcuts.R
\name{\%_\%}
\alias{\%-\%}
\alias{\%_\%}
\title{String concatenation infix operator}
\usage{
a \%_\% b

a \%-\% b
}
\arguments{
\item{a}{A string}

\item{b}{Another string}
}
\value{
a and b pasted together
}
\description{
%_% pastes two strings together (with " " as sepparator)
%-% pastes two strings together (without sepparator)
}
\examples{

"ra" \%-\% "ce" \%_\% "car"
}

