% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Metafunction to prepare data sets after reading them in}
\usage{
prepare(dat, ...)
}
\arguments{
\item{...}{}
}
\description{
Metafunction to prepare data sets after reading them in
}

