% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tex.R
\name{print_tex.Mash_table}
\alias{print_tex.Mash_table}
\title{print a Mash_table as latex}
\usage{
\method{print_tex}{Mash_table}(dat, mash_method = "row",
  insert_blank_row = (mash_method == "row"), .align = paste0("lX",
  paste(rep("X", ncol(dat[[1]]) - 1), collapse = "")),
  .include.rownames = FALSE, .floating = FALSE, .booktabs = TRUE,
  .sanitize.text.function = identity, .width = "\\\\textwidth",
  .caption = NULL, ...)
}
\arguments{
\item{dat}{an input stack table}

\item{insert_blank_row}{insert additional empty row after each row of the
tabular environment. \renewcommand{\arraystretch}{1.5}}

\item{.align}{the [pos] argument of the tabular environment. Passed on
to \code{\link{xtable}}. Please note that \code{\\newline} only
works in \code{X} columns. If you use another format the 'stacking'
will not work correclty.}

\item{.include.rownames}{passed on to \code{\link{print.xtable}}}

\item{.floating}{passed on to \code{\link{print.xtable}}}

\item{.booktabs}{passed on to \code{\link{print.xtable}}}

\item{.sanitize.text.function}{passed on to \code{\link{print.xtable}}}

\item{.width}{passed on to \code{\link{print.xtable}}}

\item{.caption}{passed on to \code{\link{xtable}}}

\item{...}{Additoinal arguments passed on to \code{\link{print.xtable}}}

\item{stack}{Stack by by \code{row} or \code{col}}

\item{.tabular.environment}{passed on to \code{\link{print.xtable}}}
}
\description{
Stacked rows are sepparated by \code{\\newline}, therefor this only works
correctly for columns that have an 'X' column type (see documentation of
the tabularx latex package). If you want each stack element in a proper
tabular row, use \code{xtable::xtable(as.data.frame(dat))} instead.
By default this uses the \code{tabularx} and \code{booktabs} latex packages.
Booktabs can be switched of, but the tabularx package is hardcoded into the
xtable settings right now as some of the formatting magic depends on it
}
