% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_prioritise.R
\name{prioritize}
\alias{prioritise}
\alias{prioritize}
\alias{prioritize.character}
\alias{prioritize.factor}
\title{Reorder character vector or levels of a factor based on priorities}
\usage{
prioritize(x, ...)

prioritise(x, ...)

\method{prioritize}{character}(x, high = character(), low = character())

\method{prioritize}{factor}(x, high = character(), low = character())
}
\arguments{
\item{x}{a character of factor vector}

\item{high}{elements to be put to the front}

\item{low}{elements to be put to the back}
}
\value{
a reordered vector
}
\description{
Shoves elements of a character or factor vector to the front.
Usefull for reordering factor levels for plotting. Issues
a warning if any elements of high or low are not present in x
}
\examples{

x <- c('d', 'e', 'z', 'y', 'n', 'b', 'c', 'a', 'x')
prioritize(x, c('a', 'b', 'c', 'applepie'), c('x', 'y', 'z'))

}

