% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quarter.R
\name{quarter}
\alias{quarter}
\title{A custom data type for Year-Quarter}
\usage{
quarter(y, q)
}
\arguments{
\item{y}{year or a character string of a format like this: 2013-Q1}

\item{q}{quarter (optional)}
}
\value{
An object of type Quarter, character
}
\description{
A custom data type for Year-Quarter
}
\examples{

quarter(2013, 3)
quarter('2013-Q3')
}
