% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quarter.R
\name{quarter_as_date}
\alias{quarter_as_date}
\title{Convert quarter to date}
\usage{
quarter_as_date(y, q)
}
\arguments{
\item{y}{Year}

\item{q}{Quarter}
}
\description{
Shorthand function for converting year - qaurter date to
}
