% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fetch_host_db2.R
\name{query_host_db2}
\alias{query_host_db2}
\title{Send Querry to Host}
\usage{
query_host_db2(q, con = RODBC::odbcConnect(dsn = gvk_secrets["dsn"], uid =
  gvk_secrets["uid"], pwd = gvk_secrets["pwd"]))
}
\description{
Send Querry to Host
}

