% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstack_table.R
\name{rstack_table}
\alias{rstack_table}
\title{Format as row-stacked tabled}
\usage{
rstack_table(line1, line2, format = "latex", insert_empty_line = FALSE,
  remove_ext = NULL, as_character = TRUE,
  format_args_line1 = list(big.mark = "~", digits = 0, format = "f"),
  format_args_line2 = list(digits = 2, format = "f"))
}
\arguments{
\item{line1}{data.frame containing the first row of grouped table}

\item{line2}{data.frame containing the second row of grouped table}

\item{insert_empty_line}{Should an empty line be inserted between grouped tables? (default FALSE for latex, FALSE for excel)}

\item{remove_ext}{Extension behind column names to remove a regex string (f.e. '_cv', '.x', etc.)}

\item{as_character}{Shoudl table cells be coerced to character? (Necessary for insert_empty_line)}
}
\description{
Format as row-stacked tabled
}

