% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_save_as.R
\name{save_as.Rstack_table_xlsx}
\alias{save_as.Rstack_table_xlsx}
\title{Title}
\usage{
save_as.Rstack_table_xlsx(dat, outfile, sheet_name = "sheet1",
  header = NULL, footer = NULL)
}
\arguments{
\item{footer}{}
}
\description{
Title
}

