% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cache.R
\name{save_cache}
\alias{save_cache}
\title{Save object to hard disk chache (of a package)}
\usage{
save_cache(..., pkg = ".", subdir)
}
\arguments{
\item{...}{R objects to save to chache dir, usually /inst/chache}

\item{subdir}{subdirectory of the chache dir to save data to}

\item{package}{the current package}
}
\description{
Save object to hard disk chache (of a package)
}
\section{Side effects}{

Saves an R object to a cache dir in the current package
}
\seealso{
\code{\link{load_cache}}.
}

