% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_test.R
\name{save_test}
\alias{save_test}
\title{Save object to hard disk chache}
\usage{
save_test(..., pkg = ".", subdir)
}
\arguments{
\item{...}{R objects to save to chache dir, usually /inst/chache}

\item{subdir}{subdirectory of the chache dir to save data to}

\item{package}{the current package}
}
\description{
Save object to hard disk chache
}
\section{Side effects}{

Saves an R object to a test_data dir in the current package
}
\seealso{
\code{\link{load_cache}}.
}

