% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_edge_pairs.R
\name{sort_edge_pairs.data.frame}
\alias{sort_edge_pairs.data.frame}
\title{Sort a data.frame where rows contain from-current-to relationships}
\usage{
\method{sort_edge_pairs}{data.frame}(x, p_name = "p", c_name = "c",
  n_name = "n", allow_partial = FALSE)
}
\arguments{
\item{p_name}{name of variable containing id of previous row}

\item{c_name}{name of variable containing id of current row}

\item{n_name}{name of variable conttaining id of next row}

\item{dat}{a data frame}
}
\description{
If the data frame row ids do not form a consecutive path, an error is thrown
}

