% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_edge_pairs.R
\name{sort_edge_pairs.default}
\alias{sort_edge_pairs.default}
\title{Sort a edge pairs into a path}
\usage{
\method{sort_edge_pairs}{default}(x, c, n, allow_partial = FALSE)
}
\arguments{
\item{c}{current node (node connecting both edges)}

\item{n}{next node}

\item{allow_partial}{If allow partial is set to FALSE, an unsortable data set
(eg not all nodes are connected) will throw an error. If set to TRUE, the
nodes are sorted as well as possible. An additonal column .id is added to
the results data.table to indicate each sorted group.}

\item{p}{revious node}
}
\value{
a sorted data.table consisting of the columns p, c and n (and .id if allow_partial = TRUE)
}
\description{
Sorts a finite number of directed edge-pairs defined by
previous, current and next node.
}
