% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_edge_pairs.R
\name{sort_edge_pairs}
\alias{sort_edge_pairs}
\title{Sort a edge pairs into a path}
\usage{
sort_edge_pairs(x, ..., allow_partial)
}
\value{
a sorted data.table consisting of the columns p, c and n (and .id if allow_partial = TRUE)
}
\description{
Sorts a finite number of directed edge-pairs defined by
previous, current and next node.
}
\details{
\code{\link{sort_edge_pairs.default}} and \code{\link{sort_edge_pairs.data.table}}
}
