% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlgen_create_table.R
\name{sqlgen_create_table}
\alias{sqlgen_create_table}
\title{Generate SQL CREATE TABLE statements}
\usage{
sqlgen_create_table(table_name, col_names, col_types, col_options = NULL,
  dialect = NULL)
}
\arguments{
\item{dialect}{}
}
\description{
Creates SQL CREATE TABLE statements from a vector of column names and
a vector of column types
}
