% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_chop.R
\name{str_chop}
\alias{str_chop}
\title{Chop up a string by position}
\usage{
str_chop(x, breaks)
}
\arguments{
\item{x}{string to chop into character vector}

\item{breaks}{positions at which to chop the string}
}
\value{
a character vector
}
\description{
Chop up a string by position
}
\examples{

str_chop('123456789a', c(1,4,6,9,100))

}
