% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_nobreak.R
\name{str_nobreak}
\alias{str_nobreak}
\title{Remove linebreaks and multiple spaces from string}
\usage{
str_nobreak(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
a character vector without linebreaks
}
\description{
Remove linebreaks and multiple spaces from string
}
