% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticktock.R
\name{tick}
\alias{tick}
\alias{tock}
\title{Tick-Tock timer}
\usage{
tick()

tock()
}
\value{
tick() displays current time and saves it to the hidden vector .tick in the global environment

tock displays current time and time difference to .tick as set by tick()
}
\description{
Convenient function to quickly measure time differences
}
\examples{
tick()
# [1] "Tick: 2015-10-28 15:45:12"
tock()
# [1] "Tock: 2015-10-28 15:45:17  - Diff:  4.1 secs"

}

