% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_match_sequence.R
\name{vec_match_seq}
\alias{vec_match_seq}
\alias{\%seq_in\%}
\title{Find position of a short vector inside a longer vector}
\usage{
vec_match_seq(sequence, target)

sequence \%seq_in\% target
}
\description{
\code{vec_match_seq} returns a vector of the positions of the matches of
a sequence inside a longer vector. The index coresponds to the where the
first element of the sequence matches the vector.
}
\details{
`%seq_in%` returns a logical vector of the matching described above
}
