% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kv_range.R
\name{within_kv_range}
\alias{within_kv_range}
\title{Check if a value is within a key-value range}
\usage{
within_kv_range(key, value, kv_range, kv_range_keycol)
}
\arguments{
\item{key}{A vector key names}

\item{value}{A vector of the same lengths with values for the specific key}

\item{kv_range}{An object of type Valid_range (see \code{\link{valid_range}})}

\item{kv_range_keycol}{The column in the Valid_range object to compare \code{key} with,}
}
\description{
Check if a value is within a key-value range
}
