% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_format.R
\name{df_format}
\alias{df_format}
\title{Format columns of a data.frame}
\usage{
df_format(dat, ...)
}
\arguments{
\item{dat}{a \code{data.frame} (or something that can be coearced with \code{as.data.frame})}

\item{...}{Key value pairst of the form classname = list(parameters to \code{format}) (see examples)}
}
\value{
a formatted data.frame
}
\description{
This function applies formatting to each column of a data.frame
that matches a specified type. For details refer to the documentation of
\link{\code{format}}
}
\examples{

tdf <- data.frame(
a = c("alpha", "beta", "ceta"),
b = c(1.230,-1124.0, 1.90),
c = c(1L, 3L, 5L),
d = as.Date(c("2009-01-12", "2009-01-12", "2009-01-12")),
t = as.POSIXct(c("2009-01-12 10:01:01", "2009-01-12 23:01:03", "2009-01-12 16:01:01"), "\%Y-\%m-\%d \%H:\%M:\%S"),
stringsAsFactors = FALSE)

df_format(tdf,
  integer = list(digits = 3, big.mark = ".", decimal.mark = ","),
  date    = list("\%m/\%d/\%y"),
  POSIXct = list("\%m/\%d/\%y \%H:\%M:\%S"))


}
\seealso{
df_format_num
}
