% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid}
\alias{is_valid}
\alias{assert_valid}
\title{Check if an object is valid}
\usage{
is_valid(x, ...)

assert_valid(dat, ...)
}
\arguments{
\item{x}{an R object}
}
\value{
logical; whether this object meets pre-defined validity conditions.
}
\description{
requires that is_valid.class is defined somewhere
}
